Confirmed that update action works through curl
Created scripts/update_cashflows.sh to run a curl request
for update on cashflows.
Commented out, 'belongs_to' and 'has_many' on cashflow and user models
because a user_id column was not present on the row I was attempting to
update. This was returning 'unprocessable_entity' prior to this change.
Eliminated a redundancy in the routes.rb file.